fix(web): better code structure, optimizations, community section, popups below buttons in desktop
