+++ web/src/pages/Home/Community/index.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import styled from "styled-components";
+import styled, { css } from "styled-components";
+import { smallScreenStyle } from "styles/smallScreenStyle";
 import { Card } from "@kleros/ui-components-library";
 import { Element } from "./Element";
 import { firstSection, secondSection } from "consts/community-elements";
@@ -19,11 +20,35 @@ const StyledSeparator = styled.hr`
 
 const Section = styled.div`
   width: 100%;
-  padding: 12px;
+  height: 64px;
+  padding: 0 32px;
   display: flex;
-  flex-direction: column;
+  flex-direction: row;
   align-items: center;
-  gap: 8px;
+  justify-content: space-between;
+
+  ${smallScreenStyle(
+    () => css`
+      flex-direction: column;
+      justify-content: center;
+      gap: 8px;
+      padding: 12px;
+      height: auto;
+    `
+  )}
+`;
+
+const TwoElementContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+  gap: 48px;
+
+  ${smallScreenStyle(
+    () => css`
+      flex-direction: column;
+      gap: 8px;
+    `
+  )}
 `;
 
 const Community = () => (
@@ -31,14 +56,17 @@ const Community = () => (
     <h1>Community</h1>
     <StyledCard>
       <Section>
-        {firstSection.map((element, i) => (
-          <Element key={i} {...element} />
-        ))}
+        <TwoElementContainer>
+          {firstSection.slice(0, 2).map((element) => (
+            <Element key={element.primaryText} {...element} />
+          ))}
+        </TwoElementContainer>
+        <Element {...firstSection[2]} />
       </Section>
       <StyledSeparator />
       <Section>
-        {secondSection.map((element, i) => (
-          <Element key={i} {...element} />
+        {secondSection.map((element) => (
+          <Element key={element.primaryText} {...element} />
         ))}
       </Section>
     </StyledCard>
